* commit 'f6f3b730a1d8046472f0d6558af8f03884d63db3 [formerly becc0918f958bd28dd36d661bf2ff06e205a08c5]': (23 commits)
  MAT-9000. Code clean up.
  MAT-9000. Changes to prevent issue of having two dialogue boxes when keyboard enter key is pressed on the delete when deleting an organization.
  MAT-8995: Clone the template node and then make changes on the clone so that the changes aren't applied to the template node in session
  MAT-8992 : Code review feedback. Remove TODO's and comments.
  MAT-9000. Bootstrap changes on Admin pages. Removed unnecessary commented out lines of code.
  default isCodeSystemVersionIncluded to true
  MAT-8995: fix issue with parent library
  MAT-8995: extract the attribute map out to a variable per code review.
  MAT-9000: Add new icon for Generate CSV in the AdministratorReports tab
  MAT-8992 : Cloning / Editing measure - messaging updates.
  MAT-8960 : Updating the validation message.
  MAT-8995: remove version from HQMF and Human Readable if includeCodeSystemVersion is false or not present.
  MAT-8996. Code clean up and re-factoring.
  MAT-8996. Fix for code system issue with adding suffix.
  MAT-8996 : Changes to display code system version conditionally in code system declaration. Removed commented code and did some clean up.
  MAT-9007. Changes to conditionally display code system in code declaration.
  MAT-9004: delete codeSystem if it is not used by other codes
  MAT-8998. Rename declaration based on code feedback.
  MAT-8998. Added code to not display version for codes Birthdate(21112-8) and Dead (419099009).
  MAT-8993: made suggested changes from code review
  ...


Former-commit-id: 740e6cdb44de27944c7e94d3f49a3dd4c1cd4658

+++ src/mat/server/CQLLibraryService.java
@@ -43,7 +43,6 @@ import mat.dao.UserDAO;
 import mat.dao.clause.CQLLibraryDAO;
 import mat.dao.clause.CQLLibrarySetDAO;
 import mat.dao.clause.CQLLibraryShareDAO;
-import mat.dao.clause.MeasureDAO;
 import mat.dao.clause.ShareLevelDAO;
 import mat.model.CQLLibraryOwnerReportDTO;
 import mat.model.CQLValueSetTransferObject;
@@ -95,10 +94,6 @@ public class CQLLibraryService extends SpringRemoteServiceServlet implements CQL
 	/** The cql library DAO. */
 	@Autowired
 	private CQLLibraryDAO cqlLibraryDAO;
-	
-	/** The measure dao. */
-	@Autowired
-	private MeasureDAO measureDAO;
 	
 	/** The cql library set DAO. */
 	@Autowired
@@ -1700,8 +1695,8 @@ public class CQLLibraryService extends SpringRemoteServiceServlet implements CQL
 	 */
 	@Override
 	public void updateUsersShare(final SaveCQLLibraryResult result) {
-		StringBuffer auditLogAdditionlInfo = new StringBuffer("CQL Library shared with ");
-		StringBuffer auditLogForModifyRemove = new StringBuffer("CQL Library shared status revoked with ");
+		StringBuilder auditLogAdditionlInfo = new StringBuilder("CQL Library shared with ");
+		StringBuilder auditLogForModifyRemove = new StringBuilder("CQL Library shared status revoked with ");
 		CQLLibraryShare cqlLibraryShare = null;
 		boolean first = true;
 		boolean firstRemove = true;
@@ -1740,14 +1735,12 @@ public class CQLLibraryService extends SpringRemoteServiceServlet implements CQL
 					
 					cqlLibraryShare.setShareLevel(sLevel);
 					cqlLibraryShareDAO.save(cqlLibraryShare);
-				} else if (!ShareLevel.MODIFY_ID.equals(dto.getShareLevel())) {
+				} else if (cqlLibraryShare != null && !ShareLevel.MODIFY_ID.equals(dto.getShareLevel())) {
 					recordRevokeShareEvent = true;
 					cqlLibraryShareDAO.delete(cqlLibraryShare.getId());
 					logger.info("Removing Sharing " + cqlLibraryShare.getCqlLibrary().getId()
 							+ " with " + user.getId()
 							+ " at level " + sLevel.getDescription());
-					System.out.println("Removing Sharing " + cqlLibraryShare.getCqlLibrary().getId()
-							+ " with " + user.getId() + " at level " + sLevel.getDescription());
 					if (!firstRemove) { //first time, don't add the comma.
 						auditLogForModifyRemove.append(", ");
 					}
@@ -1762,10 +1755,12 @@ public class CQLLibraryService extends SpringRemoteServiceServlet implements CQL
 			if (recordShareEvent && recordRevokeShareEvent) {
 				auditLogAdditionlInfo.append("\n").append(auditLogForModifyRemove);
 			} else if (recordRevokeShareEvent) {
-				auditLogAdditionlInfo = new StringBuffer(auditLogForModifyRemove);
+				auditLogAdditionlInfo = new StringBuilder(auditLogForModifyRemove);
 			}
-			cqlLibraryAuditLogDAO.recordCQLLibraryEvent(cqlLibraryShare.getCqlLibrary(),
-					"CQL Library Shared", auditLogAdditionlInfo.toString());
+			if(cqlLibraryShare != null && cqlLibraryShare.getCqlLibrary() != null) {
+				cqlLibraryAuditLogDAO.recordCQLLibraryEvent(cqlLibraryShare.getCqlLibrary(),
+						"CQL Library Shared", auditLogAdditionlInfo.toString());
+			}
 		}
 	}
